Reorganize scopegraph a bit and add ability to conduct transient scoping for tooling
