+++ graphs/scopegraph/tests/transient/transient.seru
@@ -1,0 +1,5 @@
+class SomeClass {
+    function<void> DoSomething(someParam int) {
+        /* someParam */someParam
+    }
+}
